+++ lib/core.dart
@@ -4,6 +4,7 @@ library core;
 export 'src/core/digit.dart';
 export 'src/core/equal.dart';
 export 'src/core/len.dart';
+export 'src/core/num.dart';
 export 'src/core/req.dart';
 export 'src/core/rules.dart';
 export 'src/core/trim.dart';